dev: move command preludes
